parquet_test/requirements.txt
azure-identity>=1.15.0
requests>=2.31.0
pyarrow>=14.0.0
pandas>=2.0.0
